tests: utils: rename activity_pub -> activitypub
to harmonize with the other activitypub file names
